Rename test case components #3
